// frontend/src/app/api-services/api/model/create-group-request.ts
/**
 * StudyConnect Backend API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateGroupRequest {
    name: string;
    description?: string;
    is_public: boolean;
    max_members: number;
    member_ids: Array<number>;
    admin_ids: Array<number>;
    created_by_id: number;
}